% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{BiocPkgStats-helpers}
\alias{BiocPkgStats-helpers}
\alias{typeTranslate}
\alias{typeEnglish}
\title{Internal helpers to translate values to text}
\usage{
typeTranslate(type_term)

typeEnglish(package_type)
}
\arguments{
\item{type_term}{\code{character(1)} A term to be translated obtained from
\code{biocViews::guessPackageType()}}

\item{packageType}{\code{character(1)} The term given by \code{typeTranslate} and
often used throughout the project, e.g., "software", "data-experiment",
etc.}
}
\description{
\code{typeTranslate} will take a package's \code{biocViews} terms and run
them through \code{biocViews::guessPackageType()} to obtain the package type
value. \code{typeEnglish} will take the package type value and convert it
into standard English text.
}
\keyword{internal}
