% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Create a report a given set of packages}
\usage{
generateReport(packages, gh_org, since_date, template, overwrite = FALSE)
}
\arguments{
\item{packages}{\code{character()} A vector of valid package names}

\item{gh_org}{\code{character(1)} The GitHub organization from which to read
issue and commit data from.}

\item{since_date}{\code{character(1)} The date from when to start looking at
commit and issue history. This should be specified in the year, month, and
day format, 'YYYY-MM-DD'.}

\item{template}{\code{character(1)} The file location of the template, this
defaults to the internal template}

\item{overwrite}{\code{logical(1)} Whether to overwrite an existing rendered
product, i.e., a runnable RMarkdown document.}
}
\description{
Create a report a given set of packages
}
\examples{
if (interactive()) {

generateReport(
    c(
        "MultiAssayExperiment", "cBioPortalData", "SingleCellMultiModal"
    ),
    gh_org = "waldronlab",
    since_date = "2021-05-01",
    template = "~/test/Pkg_report_template.Rmd",
    overwrite = TRUE
)

generateReport(
    "RaggedExperiment",
    gh_org = "Bioconductor",
    since_date = "2021-05-01",
    template = "~/test/Pkg_report_template.Rmd",
    overwrite = TRUE
)

}

}
