% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTable.R
\name{generateTable}
\alias{generateTable}
\title{Create a table of package metrics}
\usage{
generateTable(packages, gh_org, since_date)
}
\arguments{
\item{packages}{\code{character()} A vector of valid package names that are
installed}

\item{gh_org}{\code{character()} The GitHub organization from which to read issue
and commit data from. It can be either a scalar character to be recycled to
the length of packages or a vector the length of \code{packages} for instances
where packages are hosted under different GitHub organizations.}

\item{since_date}{\code{character()} The date from when to start looking at
commit and issue history. This should be specified in the year, month, and
day format, 'YYYY-MM-DD'. It can be a vector of dates that match the
length \code{packages}.}
}
\value{
A \code{data.frame} of metrics, see the \code{details} section for specifics
}
\description{
The function compiles statistics for Bioconductor packages that
have GitHub repositories. It uses functionality in \code{BiocPkgTools} to
extract commit and issue history. A token is required to access the
GitHub commit and issue history. See the \code{.token} argument in \code{?gh::gh}
for details on its use.
}
\details{
Note that pull requests are a type of "issue" on GitHub; therefore, issues
and pull requests are combined when reporting the number of closed issues
since a specific date. For more information on how the queries are preformed,
see the GitHub API documentation for the \verb{/repos/\{owner\}/\{repo\}/issues}
endpoint at \url{https://docs.github.com/en/rest/issues} and the \code{gh} R
package. The following list describes each column in the returned
\code{data.frame}:

\itemize{
\item{download.rank:} The percentile rank of the package in its
respective repository, e.g., software, data-experiment etc.
\item{avg.downloads:} The average number of monthly downloads from
distinct IP addresses
\item{num.revdeps:} The number of "all" reverse dependencies including
"Depends", "Imports", "LinkingTo", "Suggests", and "Enhances"
\item{issues.since:} The number of closed issues (including pull
requests) since the date given
\item{commits.since:} The number of commits since the given date
}
}
\examples{
if (interactive()) {

generateTable(
    packages = c(
        "MultiAssayExperiment", "cBioPortalData", "SingleCellMultiModal"
    ),
    gh_org = "waldronlab",
    since_date = "2019-05-01"
)

}
}
