% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTable.R
\name{generateTable}
\alias{generateTable}
\title{Create a table of package metrics}
\usage{
generateTable(packages, gh_org, since_date)
}
\arguments{
\item{packages}{\code{character()} A vector of valid package names that are
installed}

\item{gh_org}{\code{character()} The GitHub organization from which to read issue
and commit data from. It can be either a scalar character to be recycled to
the length of packages or a vector the length of \code{packages} for instances
where packages are hosted under different GitHub organizations.}

\item{since_date}{\code{character()} The date from when to start looking at
commit and issue history. This should be specified in the year, month, and
day format, 'YYYY-MM-DD'. It can be a vector of dates that match the
length \code{packages}.}
}
\value{
A data.frame of metrics including download rank, average number of
monthly downloads, number of reverse dependencies, issues closed and
commits done since the given date.
}
\description{
The function compiles statistics for Bioconductor packages that
have GitHub repositories. It uses functionality in \code{BiocPkgTools} to
extract commit and issue history. A token is required to access the
GitHub commit and issue history. See the \code{.token} argument in \code{?gh::gh}
for details on its use.
}
\examples{
if (interactive()) {

generateTable(
    packages = c(
        "MultiAssayExperiment", "cBioPortalData", "SingleCellMultiModal"
    ),
    gh_org = "waldronlab",
    since_date = "2019-05-01"
)

}
}
