% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportSiteStats.R
\name{supportSiteStats}
\alias{supportSiteStats}
\alias{print.bioc_support_stats}
\title{Gather support site statistics for a given period}
\usage{
supportSiteStats(from = "2021-01-01", to = "2021-12-31")

\method{print}{bioc_support_stats}(x, ...)
}
\arguments{
\item{from}{\code{character(1)} date in ISO 8601 format ("YYYY-MM-DD")}

\item{to}{\code{character(1)} date in ISO 8601 format ("YYYY-MM-DD")}
}
\value{
a \code{bioc_support_stats} list class with components userdiff,
toplevdiff, questdiff, and respdiff
}
\description{
Gather support site statistics for a given period
}
\examples{

supportSiteStats()

}
\author{
Vincent J. Carey
}
